# Reduce allocations in models.(*Point)