--- a/models/flat_tags.go
+++ b/models/flat_tags.go
@@ -20,6 +20,7 @@ func flatTagsGetFromPool() *flatTags {
 // flatTagsPutIntoPool resets a flatTags instance and puts it into its pool.
 func flatTagsPutIntoPool(x *flatTags) {
 	x.Data = x.Data[:0]
+	x.Sorted = true
 	flatTagsPool.Put(x)
 }
 
@@ -31,6 +32,7 @@ type flatTag struct {
 // flatTags holds a slice of flatTag instances. It is a lower-overhead
 // alternative to map[string]string that is used elsewhere in this package.
 type flatTags struct {
+	Sorted bool
 	Data []flatTag
 }
 
@@ -51,6 +53,12 @@ func (a *flatTags) Get(i int) ([]byte, []byte) {
 	return x.Key, x.Val
 }
 
+// IsSorted is a simple predicate. Use it to skip unnecessary calls to
+// InsertionSort.
+func (a *flatTags) IsSorted() bool {
+	return a.Sorted
+}
+
 // InsertionSort is an in-place zero-overhead sort. It should be used instead
 // of sort.Sort, to avoid heap allocations. It is less efficient for larger
 // sets of items.
@@ -60,6 +68,7 @@ func (a *flatTags) InsertionSort() {
 			a.Swap(j, j-1)
 		}
 	}
+	a.Sorted = true
 }
 
 // Append appends a new key, value pair to this flatTags instance.
@@ -68,4 +77,9 @@ func (a *flatTags) Append(k, v []byte) {
 		Key: k,
 		Val: v,
 	})
+	if a.Sorted && len(a.Data) >= 2 {
+		i := len(a.Data) - 2
+		j := len(a.Data) - 1
+		a.Sorted = a.Less(i, j)
+	}
 }