--- a/models/simple_unsafe.go
+++ b/models/simple_unsafe.go
@@ -18,3 +18,18 @@ func unsafeBytesToString(in []byte) string {
 	s := *(*string)(unsafe.Pointer(&dst))
 	return s
 }
+
+// unsafeStringToBytes converts a string to a []byte without a heap allocation.
+//
+// It is unsafe, and is intended to prepare input to short-lived functions
+// that require byte slices.
+func unsafeStringToBytes(in string) []byte {
+	src := *(*reflect.StringHeader)(unsafe.Pointer(&in))
+	dst := reflect.SliceHeader{
+		Data: src.Data,
+		Len:  src.Len,
+		Cap:  src.Len,
+	}
+	s := *(*[]byte)(unsafe.Pointer(&dst))
+	return s
+}