###### Required for all non-trivial PRs
- [*] Rebased/mergable
- [*] Tests pass
- [*] Sign [CLA](https://influxdata.com/community/cla/) (if not already signed)

###### Summary

This PR reduces GC pressure by reducing allocations during various `models.(*Point)` operations.
It does not change any public function signatures.

###### Changes

+ Use `flatTags` instead of `map[string]string` as a bookeeping data structure during `models.(*Point).HashKey`. This includes tests and a global pool.
+ Implement `[]byte -> string` and `string -> []byte` casts, and use them (carefully) in cases where the Go compiler fails to keep the value on the stack.
+ Use a slice of struct to store escape codes, not a map (due to iteration speed differences).

###### `models` Benchmarks before and after

```
After:
BenchmarkMarshal-4                       2000000              1520 ns/op              64 B/op          1 allocs/op
BenchmarkParsePointNoTags-4              5000000               686 ns/op          33.50 MB/s         160 B/op          2 allocs/op*
BenchmarkParsePointWithPrecisionN-4      5000000               665 ns/op          34.54 MB/s         160 B/op          2 allocs/op*
BenchmarkParsePointWithPrecisionU-4      5000000               681 ns/op          33.77 MB/s         160 B/op          2 allocs/op*
BenchmarkParsePointsTagsSorted2-4        3000000              1023 ns/op          49.84 MB/s         160 B/op          2 allocs/op*
BenchmarkParsePointsTagsSorted5-4        2000000              1311 ns/op          63.28 MB/s         160 B/op          2 allocs/op*
BenchmarkParsePointsTagsSorted10-4       1000000              2019 ns/op          70.82 MB/s         160 B/op          2 allocs/op*
BenchmarkParsePointsTagsUnSorted2-4      2000000              1220 ns/op          41.77 MB/s         192 B/op          3 allocs/op*
BenchmarkParsePointsTagsUnSorted5-4      2000000              1862 ns/op          44.56 MB/s         224 B/op          3 allocs/op*
BenchmarkParsePointsTagsUnSorted10-4     1000000              3304 ns/op          43.28 MB/s         288 B/op          3 allocs/op*
BenchmarkParseKey-4                      1000000              2829 ns/op            1030 B/op         24 allocs/op

Before:
BenchmarkMarshal-4                       1000000              3289 ns/op             560 B/op         13 allocs/op
BenchmarkParsePointNoTags-4              3000000               791 ns/op          29.07 MB/s         208 B/op          4 allocs/op
BenchmarkParsePointWithPrecisionN-4      5000000               779 ns/op          29.50 MB/s         208 B/op          4 allocs/op
BenchmarkParsePointWithPrecisionU-4      5000000               775 ns/op          29.66 MB/s         208 B/op          4 allocs/op
BenchmarkParsePointsTagsSorted2-4        3000000              1087 ns/op          46.91 MB/s         240 B/op          4 allocs/op
BenchmarkParsePointsTagsSorted5-4        2000000              1446 ns/op          57.39 MB/s         272 B/op          4 allocs/op
BenchmarkParsePointsTagsSorted10-4       1000000              2284 ns/op          62.59 MB/s         320 B/op          4 allocs/op
BenchmarkParsePointsTagsUnSorted2-4      2000000              1393 ns/op          36.59 MB/s         272 B/op          5 allocs/op
BenchmarkParsePointsTagsUnSorted5-4      1000000              2143 ns/op          38.72 MB/s         336 B/op          5 allocs/op
BenchmarkParsePointsTagsUnSorted10-4     1000000              3669 ns/op          38.97 MB/s         448 B/op          5 allocs/op
BenchmarkParseKey-4                      1000000              2818 ns/op            1030 B/op         24 allocs/op

[*] The benchmarks were updated to remove one spurious heap allocation per iteration.
```