@jwilder 2016-08-15
The `flatTags` struct is very similar to the changes in #7138.  See: https://github.com/influxdata/influxdb/pull/7138/files#diff-fe01dfb5f49b26432bd2d62716a991a5R1395