@rw 2016-08-15
@jwilder I see, the `Tags` changes have some similar aspects. That PR is rather involved, when do you think it will be merged?