@jsternberg 2016-09-02
It looks like #7138 was already merged. @rw is this issue still needed or will #7138 resolve this? Thanks.