@jsternberg 2016-09-14
I'm closing this. If it's still needed, please rebase and comment and I can reopen this.